[Move/Examples] Switch to datatest
## Description
Use `datatest-stable` to find all the Move examples we might want to
build and test, instead of stashing this away in a rust test.

## Test plan

```
sui$ cargo nextest run -p sui-framework-tests --test move_tests
```
