+++ .github/workflows/rust.yml
@@ -262,7 +262,7 @@ jobs:
       - uses: taiki-e/install-action@nextest
       - name: Run move tests
         run: |
-          cargo nextest run -p sui-framework-tests -- unit_tests::
+          cargo nextest run -p sui-framework-tests --test move_tests
 
   # # Disabled
   # rosetta-validation:
