+++ crates/sui-framework-tests/src/lib.rs
@@ -2,7 +2,4 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #[cfg(test)]
-mod unit_tests;
-
-#[cfg(test)]
 mod metered_verifier;
